+++ gulpfile.js
@@ -4,7 +4,7 @@ const config = require(__dirname+'/lib/misc/config.js')
 	, { Binary } = require('mongodb')
 	, Permission = require(__dirname+'/lib/permission/permission.js')
 	, { Permissions } = require(__dirname+'/lib/permission/permissions.js')
-	, { hcaptcha, google } = require(__dirname+'/configs/secrets.js')
+	, { hcaptcha, google, yandex } = require(__dirname+'/configs/secrets.js')
 	, gulp = require('gulp')
 //	, pugRuntime = require('pug-runtime/build')
 	, fs = require('fs-extra')
@@ -301,6 +301,11 @@ async function css() {
 				captchaHeight = 200;
 				captchaWidth = 200;
 				break;
+			case 'yandex':
+				bypassHeight = 500;
+				captchaWidth = 300;
+				captchaHeight = 82;
+				break;
 			case 'grid':
 			case 'grid2':
 				bypassHeight = 330;
@@ -428,6 +433,7 @@ async function custompages() {
 			postFilesSize: formatSize(config.get.globalLimits.postFilesSize.max),
 			googleRecaptchaSiteKey: google.siteKey,
 			hcaptchaSiteKey: hcaptcha.siteKey,
+			yandexSiteKey: yandex.siteKey,
 			globalAnnouncement: config.get.globalAnnouncement,
 			captchaOptions: config.get.captchaOptions,
 			commit,